Split settings activity (#61)
* splitting settings activity

* removed duplication

* auditing the app with ai

* standarizing default system label
