+++ app/src/main/java/com/rama/mako/activities/AboutActivity.kt
@@ -11,11 +11,6 @@ class AboutActivity : CsActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        @Suppress("DEPRECATION")
-        window.decorView.systemUiVisibility =
-            View.SYSTEM_UI_FLAG_LAYOUT_STABLE or
-                    View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-
         setContentView(R.layout.view_about)
 
         val root = findViewById<View>(android.R.id.content)
